schedule volume resource (#161)
* add volume_map into protobuf

* node management in rpc layer

* node management in cluster and store

* container operation with volume on protobuf

* cluster interface and rpc adapter

* calcium can UpdateNodeResource with volume map: create / dissociate /
remove

* implement multiple fragments scheduling algorithm

* share host struct to distribute resource

* bugfix and pass test

* specify int64 for ResourceMap

* introduce VolumePlan type, create volumes using scheduled plans

* store.AddNode interface: introduce parameter object

* engine handles auto volume

* unittest for volume distribution algorithm

* createContainerMessage contains volume plan

* refactor: rename, minors

* realloc volume schedule

* typo, bugfix and minors

* replenish types/node_test

* keep the same device binding as existing container's(vm's) when realloc

* unittest for realloc

* soothe lint

* get container with volumes and volume plan

* unittest covers realloc more

* realloc hard volume, like /tmp:/tmp

* engine knows whether volume changed by realloc request

* delete VolumeUsed in NodeInfo: useless

* remedy CompareStringSlice comments

+++ engine/docker/docker.go
@@ -83,7 +83,7 @@ func (e *Engine) Info(ctx context.Context) (*enginetypes.Info, error) {
 }
 
 // ResourceValidate validate resource usage
-func (e *Engine) ResourceValidate(ctx context.Context, cpu float64, cpumap map[string]int, memory, storage int64) error {
+func (e *Engine) ResourceValidate(ctx context.Context, cpu float64, cpumap map[string]int64, memory, storage int64) error {
 	// TODO list all containers, calcuate resource
 	return nil
 }
