+++ engine/engine.go
@@ -48,5 +48,5 @@ type API interface {
 	VirtualizationUpdateResource(ctx context.Context, ID string, opts *enginetypes.VirtualizationResource) error
 	VirtualizationCopyFrom(ctx context.Context, ID, path string) (io.ReadCloser, string, error)
 
-	ResourceValidate(ctx context.Context, cpu float64, cpumap map[string]int, memory, storage int64) error
+	ResourceValidate(ctx context.Context, cpu float64, cpumap map[string]int64, memory, storage int64) error
 }
